requirements.txt
openai>=1.0.0
python-dotenv
pyttsx3
SpeechRecognition
pyaudio
pygame
pillow
gTTS
